csp_bridge: Move csp_bridge_start to arch
csp_bridge_start() and csp_bridge() are architecture dependent,
meaning that they must know the thread implementation of calling
architecture.

Make sure to call csp_<arch>_thread_create() so that we can further
reduce the users of csp_thread_create().
