+++ src/arch/macosx/meson.build
@@ -1,4 +1,5 @@
 csp_sources += files([
+  'csp_bridge.c',
   'csp_clock.c',
   'csp_queue.c',
   'csp_semaphore.c',
