+++ src/arch/posix/meson.build
@@ -1,4 +1,5 @@
 csp_sources += files([
+	'csp_bridge.c',
 	'csp_queue.c',
 	'csp_semaphore.c',
 	'csp_system.c',
@@ -9,4 +10,3 @@ csp_sources += files([
 ])
 
 csp_deps += dependency('threads')
-
